% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quboleCommand.R
\docType{data}
\name{HiveCommand}
\alias{HiveCommand}
\title{Uses given authtoken to authenticate against the given api_url
 and returns and instance of qubole HiveCommand
 \url{http://qubole.com}
 \email{sdk-dev@qubole.com}}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fb8a29cc340> 
 - attr(*, "name")= chr "HiveCommand_generator"
}}
\usage{
HiveCommand
}
\arguments{
\item{api_token}{A String}

\item{poll_interval}{A Number}
}
\value{
the qubole HiveCommand instance
}
\description{
Uses given authtoken to authenticate against the given api_url
 and returns and instance of qubole HiveCommand
 \url{http://qubole.com}
 \email{sdk-dev@qubole.com}
}
\examples{
\dontrun{qubolehivecommand = HiveCommand("apitoken","pollinterval")}
}
\keyword{datasets}

