% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quboleCommand.R
\docType{data}
\name{PrestoCommand}
\alias{PrestoCommand}
\title{Uses given authtoken to authenticate against the given api_url
 and returns and instance of qubole PrestoCommand
 \url{http://qubole.com}
 \email{sdk-dev@qubole.com}}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fb8a29de068> 
 - attr(*, "name")= chr "PrestoCommand_generator"
}}
\usage{
PrestoCommand
}
\arguments{
\item{api_token}{A String}

\item{poll_interval}{A Number}
}
\value{
the qubole PrestoCommand instance
}
\description{
Uses given authtoken to authenticate against the given api_url
 and returns and instance of qubole PrestoCommand
 \url{http://qubole.com}
 \email{sdk-dev@qubole.com}
}
\examples{
\dontrun{quboleprestocommand = PrestoCommand("apitoken")}
}
\keyword{datasets}

