% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qds.sdk.R-package.R
\docType{package}
\name{qds.sdk.R}
\alias{qds.sdk.R}
\alias{qds.sdk.R-package}
\title{qds.sdk.R}
\description{
The qds.sdk.R package performs a simple call to qds-sdk-py to retriev results of a HiveCommand.
}
\author{
Karandeep Johar: \email{karandeepj@qubole.com} and
Dinesh Rathi: \email{dineshr@qubole.com}
}

