% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quboleHiveCommand.R
\name{quboleHiveCommand}
\alias{quboleHiveCommand}
\title{calls hivecommand_from_r and return the result}
\usage{
quboleHiveCommand(...)
}
\arguments{
\item{query}{A String}

\item{script_location}{A String}

\item{macros}{A String}

\item{tags}{A String}

\item{sample_size}{A Number}

\item{cluster_label}{A String}

\item{notify}{A Boolean}

\item{name}{A String}

\item{poll_interval}{A Number}

\item{api_token}{A String}

\item{api_url}{A String}

\item{api_version}{A Number}

\item{skip_ssl_cert_check}{A Boolean}

\item{verbose}{A Boolean}

\item{chatty}{A Boolean}
}
\value{
the results of the query
}
\description{
\url{http://qubole.com}
 \email{karandeepj@qubole.com}
}
\examples{
quboleHiveCommand("show tables")
}

