% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quboleCommand.R
\docType{data}
\name{SparkCommand}
\alias{SparkCommand}
\title{Uses given authtoken to authenticate against the given api_url
 and returns and instance of qubole SparkCommand
 \url{http://qubole.com}
 \email{sdk-dev@qubole.com}}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fb8a29f0590> 
 - attr(*, "name")= chr "SparkCommand_generator"
}}
\usage{
SparkCommand
}
\arguments{
\item{api_token}{A String}

\item{poll_interval}{A Number}
}
\value{
the qubole SparkCommand instance
}
\description{
Uses given authtoken to authenticate against the given api_url
 and returns and instance of qubole SparkCommand
 \url{http://qubole.com}
 \email{sdk-dev@qubole.com}
}
\examples{
\dontrun{qubolesparkcommand = SparkCommand("apitoken")}
}
\keyword{datasets}

