% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quboleHiveCommand.R
\name{quboleHiveCommand}
\alias{quboleHiveCommand}
\title{calls hivecommand_from_r  in /exec/r-python.py and return the result}
\usage{
quboleHiveCommand(query = NULL, poll_interval = NULL, sample_size = NULL,
  macros = NULL, tags = NULL, cluster_label = NULL, notify = NULL,
  name = NULL, api_token = NULL)
}
\arguments{
\item{query}{A String}

\item{poll_interval}{A Number}

\item{sample_size}{A Number}

\item{macros}{A String}

\item{tags}{A String}

\item{cluster_label}{A String}

\item{notify}{A Boolean}

\item{name}{A String}

\item{api_token}{A String}
}
\value{
the results of the query
}
\description{
\url{http://qubole.com}
 \email{sdk-dev@qubole.com}
}
\examples{
\dontrun{results = quboleHiveCommand("show tables", poll_interval = 5)}
}

